Add bard support
