+++ fflogs_rotation/job_data/potencies/physical_ranged/Bard/7_01-Bard.csv
@@ -5,7 +5,7 @@ Bard,7409,Refulgent Arrow,100,,280,direct,,,,,,
 Bard,8,Shot,100,,80,auto,,,,,,
 Bard,117,Rain of Death,100,,100,direct,,,,,,
 Bard,16494,Shadowbite,100,,170,direct,,,,,,
-Bard,16496,Apex Arrow,100,gauge_90,550,direct,,,,,,
+Bard,16496,Apex Arrow,100,gauge_90,540,direct,,,,,,
 Bard,16496,Apex Arrow,100,gauge_100,600,direct,,,,,,
 Bard,3558,Empyreal Arrow,100,,260,direct,,,,,,
 Bard,7404,Pitch Perfect,100,pp1,100,direct,,,,,,
